+++ src/main/java/com/nexters/phochak/controller/LikesController.java
@@ -3,9 +3,9 @@ package com.nexters.phochak.controller;
 import com.nexters.phochak.auth.UserContext;
 import com.nexters.phochak.auth.annotation.Auth;
 import com.nexters.phochak.dto.response.CommonResponse;
+import com.nexters.phochak.service.LikesService;
 import com.nexters.phochak.exception.PhochakException;
 import com.nexters.phochak.exception.ResCode;
-import com.nexters.phochak.service.PhochakService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -19,14 +19,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("v1/post/{postId}/likes")
 public class LikesController {
 
-    private final PhochakService phochakService;
+    private final LikesService likesService;
 
     @Auth
     @PostMapping
     public CommonResponse<Void> addPhochak(@PathVariable Long postId) {
         Long userId = UserContext.getContext();
         try {
-            phochakService.addPhochak(userId, postId);
+            likesService.addPhochak(userId, postId);
         } catch (DataIntegrityViolationException e) {
             throw new PhochakException(ResCode.ALREADY_PHOCHAKED);
         }
@@ -37,7 +37,7 @@ public class LikesController {
     @DeleteMapping
     public CommonResponse<Void> cancelPhochak(@PathVariable Long postId) {
         Long userId = UserContext.getContext();
-        phochakService.cancelPhochak(userId, postId);
+        likesService.cancelPhochak(userId, postId);
         return new CommonResponse<>();
     }
 }
