--- src/main/java/com/nexters/phochak/service/PhochakService.java
@@ -1,7 +1,0 @@
-package com.nexters.phochak.service;
-
-public interface PhochakService {
-    void addPhochak(Long userId, Long postId);
-
-    void cancelPhochak(Long userId, Long postId);
-}
